LU-5478 osc: get rid of obd_* typedefs
We have a bunch of typedefs for common things that made no sense
and hid the actual type from plain view.
Replace them with proper uXX or sXX types.
Exception is in lustre_idl.h and lustre_ioctl.h where
they are replaced with __uXX and __sXX to be able to be included
in userspace

Also fix valid flags to u64.

patch 6 in series: modify osc/osp

+++ lustre/include/obd_class.h
@@ -322,7 +322,7 @@ struct inode;
 void obdo_from_la(struct obdo *dst, const struct lu_attr *la, __u64 valid);
 void la_from_obdo(struct lu_attr *la, const struct obdo *dst, u32 valid);
 
-void obdo_cpy_md(struct obdo *dst, const struct obdo *src, u32 valid);
+void obdo_cpy_md(struct obdo *dst, const struct obdo *src, u64 valid);
 void obdo_to_ioobj(const struct obdo *oa, struct obd_ioobj *ioobj);
 
 #define OBT(dev)        (dev)->obd_type
